Fix entity baseline problems
The previous implementation of entity-baselines was broken which resulted
in wrong property values being sent out & stored on the entities (#25).

It also fired multiple updates for values that didn't match the default
value set in the baseline (#24) when an entity was created.

Now there is only one update right after the entity is created and the
values should be correct.