entities: fix Destroy() not being called in some cases (#139)
when a player disconnects and is replaced by a bot, the bot may have the
same entity-ID as the player.
also the player's entity may not have been destroyed, since the
PVS-leave update was never sent.
for this case we need to make sure old entities are destoryed before
calling ReadEnterPVS().