events: delay dispatching of ItemEquip (#142)
If a bot takes over a player the item_equip event may be sent out before
the new (bot) player-entity is created.
To avoid ItemEquip.Player == nil we need to delay the event to the end
of the tick.

This commit also changes how events are delayed. Now the whole event handler
function is executed at the end of the tick, not just the dispatching.